chore: separando classes de ibscbs em arquivos diferentes
