+++ src/OpenAC.Net.NFSe/Nota/IBSCBSDocumentoOutro.cs
@@ -1,0 +1,10 @@
+using OpenAC.Net.Core.Generics;
+
+namespace OpenAC.Net.NFSe.Nota;
+
+public sealed class IBSCBSDocumentoOutro : GenericClone<IBSCBSDocumentoOutro>
+{
+    public string? NumeroDocumento { get; set; }
+
+    public string? DescricaoDocumento { get; set; }
+}
