+++ src/OpenAC.Net.NFSe/Nota/IBSCBSTotalIBSUF.cs
@@ -1,0 +1,10 @@
+using OpenAC.Net.Core.Generics;
+
+namespace OpenAC.Net.NFSe.Nota;
+
+public sealed class IBSCBSTotalIBSUF : GenericClone<IBSCBSTotalIBSUF>
+{
+    public decimal ValorDiferimento { get; set; }
+
+    public decimal ValorIBSUF { get; set; }
+}
