+++ src/OpenAC.Net.NFSe/Nota/IBSCBSValoresFed.cs
@@ -1,0 +1,12 @@
+using OpenAC.Net.Core.Generics;
+
+namespace OpenAC.Net.NFSe.Nota;
+
+public sealed class IBSCBSValoresFed : GenericClone<IBSCBSValoresFed>
+{
+    public decimal PercentualCBS { get; set; }
+
+    public decimal PercentualReducaoAliquotaCBS { get; set; }
+
+    public decimal PercentualAliquotaEfetivaCBS { get; set; }
+}
