+++ src/OpenAC.Net.NFSe/Nota/InfoTributosIBSCBS.cs
@@ -1,0 +1,13 @@
+using OpenAC.Net.Core.Generics;
+
+namespace OpenAC.Net.NFSe.Nota;
+
+public sealed class InfoTributosIBSCBS : GenericClone<InfoTributosIBSCBS>
+{
+    public InfoTributosIBSCBS()
+    {
+        SituacaoClassificacao = new InfoTributosSitClass();
+    }
+
+    public InfoTributosSitClass SituacaoClassificacao { get; set; }
+}
