+++ src/OpenAC.Net.NFSe/Nota/InfoTributosSitClass.cs
@@ -1,0 +1,10 @@
+using OpenAC.Net.Core.Generics;
+
+namespace OpenAC.Net.NFSe.Nota;
+
+public sealed class InfoTributosSitClass : GenericClone<InfoTributosSitClass>
+{
+    public string? CodigoSituacaoTributaria { get; set; }
+
+    public string? CodigoClassificacaoTributaria { get; set; }
+}
